Move popover to components
